requirements.txt
click>=7
ffmpeg-python
https://github.com/pyannote/pyannote-audio/tarball/develop
opentimelineio==0.11.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=1.6